Sets default root directory for COCO data
Updates the default root directory for COCO data to '/root/lanyun-tmp'.

This change simplifies usage by allowing users to omit the --root argument
when using the default directory, while still providing the option to
override it. This avoids unnecessary specification of the root directory
for users who are working within the default environment.